+++ examples_rclcpp/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package examples_rclcpp
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (2019-10-07)
+------------------
+* Added example node for action server and client
+* Contributors: SungHyeon Oh, Darby Lim
+
 1.0.0 (2019-08-05)
 ------------------
 * Changed code for new style guide