+++ examples_rclpy/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package examples_rclpy
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (2019-10-07)
+------------------
+* None
+
 1.0.0 (2019-08-05)
 ------------------
 * Changed code for new style guide and PEP-8